requirements.txt
flask==3.0.3
flask-login==0.6.3
psycopg2-binary==2.9.9
sqlalchemy==2.0.29
stripe==9.5.0
openai==1.30.1
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.131.0
python-dotenv==1.0.1
werkzeug==3.0.3
gunicorn==22.0.0
APScheduler==3.10.4
requests==2.32.3
